+++ src/level/tile/tile.hpp
@@ -32,11 +32,11 @@ class Tile
 
     virtual bool may_pass(Level* level, int x, int y, Entity* e);
     virtual int get_light_radius(Level* level, int x, int y);
-    virtual void hurt(Level* level, int x, int y, Mob* source, int damage, int attack_dir);
+    virtual void hurt(Level* level, int x, int y, Mob* source, uint8_t damage, uint8_t attack_dir);
     virtual void bumped_into(Level* level, int xt, int yt, Entity* entity);
-    virtual void tick(Level* level, int xt, int yt, int modifier);
+    virtual void tick(Level* level, int xt, int yt, uint8_t modifier);
     virtual void stepped_on(Level* level, int xt, int yt, Entity* entity);
-    virtual bool interact(Level* level, int xt, int yt, Player* player, Item* item, int attack_dir);
+    virtual bool interact(Level* level, int xt, int yt, Player* player, Item* item, uint8_t attack_dir);
     virtual bool connects_to_liquid();
 };
 #endif
